requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-multipart>=0.0.6
supabase>=2.3.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
psycopg[binary]>=3.2.0
tiktoken>=0.7.0
apscheduler>=3.10.0
itsdangerous>=2.1.0
cryptography>=42.0.0
openai>=1.12.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
tzdata>=2024.1
svix>=1.16.0
